polymarket-copy-trading/requirements.txt
py-clob-client>=0.18.0
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0
aiohttp>=3.9
aiosqlite>=0.20
click>=8.1
rich>=13.0
structlog>=24.0

# Dev dependencies
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
ruff>=0.4
